+++ HaloOnlineTagTool/TagStructures/BeamSystem.cs
@@ -13,7 +13,7 @@ namespace HaloOnlineTagTool.TagStructures
 	public class BeamSystem
 	{
 		public List<BeamSystemBlock> BeamSystem2;
-		public float Unknown;
+		public uint Unknown;
 
 		[TagStructure(Size = 0x208)]
 		public class BeamSystemBlock
@@ -27,101 +27,101 @@ namespace HaloOnlineTagTool.TagStructures
 			public sbyte Unknown3;
 			public sbyte Unknown4;
 			public sbyte Unknown5;
-			public float Unknown6;
+			public uint Unknown6;
 			public int Unknown7;
-			public float Unknown8;
-			public float Unknown9;
-			public float Unknown10;
-			public float Unknown11;
-			public float Unknown12;
-			public float Unknown13;
-			public float Unknown14;
-			public float Unknown15;
-			public float Unknown16;
+			public uint Unknown8;
+			public uint Unknown9;
+			public uint Unknown10;
+			public uint Unknown11;
+			public uint Unknown12;
+			public uint Unknown13;
+			public uint Unknown14;
+			public uint Unknown15;
+			public uint Unknown16;
 			public sbyte Input;
 			public sbyte InputRange;
 			public OutputKindValue OutputKind;
 			public sbyte Output;
 			public byte[] Unknown17;
-			public float Unknown18;
-			public float Unknown19;
+			public uint Unknown18;
+			public uint Unknown19;
 			public sbyte Input2;
 			public sbyte InputRange2;
 			public OutputKindValue2 OutputKind2;
 			public sbyte Output2;
 			public byte[] Unknown20;
-			public float Unknown21;
-			public float Unknown22;
+			public uint Unknown21;
+			public uint Unknown22;
 			public sbyte Input3;
 			public sbyte InputRange3;
 			public OutputKindValue3 OutputKind3;
 			public sbyte Output3;
 			public byte[] Unknown23;
-			public float Unknown24;
-			public float Unknown25;
+			public uint Unknown24;
+			public uint Unknown25;
 			public sbyte Input4;
 			public sbyte InputRange4;
 			public OutputKindValue4 OutputKind4;
 			public sbyte Output4;
 			public byte[] Unknown26;
-			public float Unknown27;
-			public float Unknown28;
-			public float Unknown29;
-			public float Unknown30;
-			public float Unknown31;
-			public float Unknown32;
+			public uint Unknown27;
+			public uint Unknown28;
+			public uint Unknown29;
+			public uint Unknown30;
+			public uint Unknown31;
+			public uint Unknown32;
 			public sbyte Input5;
 			public sbyte InputRange5;
 			public OutputKindValue5 OutputKind5;
 			public sbyte Output5;
 			public byte[] Unknown33;
-			public float Unknown34;
-			public float Unknown35;
+			public uint Unknown34;
+			public uint Unknown35;
 			public sbyte Input6;
 			public sbyte InputRange6;
 			public OutputKindValue6 OutputKind6;
 			public sbyte Output6;
 			public byte[] Unknown36;
-			public float Unknown37;
-			public float Unknown38;
+			public uint Unknown37;
+			public uint Unknown38;
 			public sbyte Input7;
 			public sbyte InputRange7;
 			public OutputKindValue7 OutputKind7;
 			public sbyte Output7;
 			public byte[] Unknown39;
-			public float Unknown40;
-			public float Unknown41;
+			public uint Unknown40;
+			public uint Unknown41;
 			public sbyte Input8;
 			public sbyte InputRange8;
 			public OutputKindValue8 OutputKind8;
 			public sbyte Output8;
 			public byte[] Unknown42;
-			public float Unknown43;
-			public float Unknown44;
+			public uint Unknown43;
+			public uint Unknown44;
 			public sbyte Input9;
 			public sbyte InputRange9;
 			public OutputKindValue9 OutputKind9;
 			public sbyte Output9;
 			public byte[] Unknown45;
-			public float Unknown46;
-			public float Unknown47;
+			public uint Unknown46;
+			public uint Unknown47;
 			public sbyte Input10;
 			public sbyte InputRange10;
 			public OutputKindValue10 OutputKind10;
 			public sbyte Output10;
 			public byte[] Unknown48;
-			public float Unknown49;
-			public float Unknown50;
+			public uint Unknown49;
+			public uint Unknown50;
 			public sbyte Input11;
 			public sbyte InputRange11;
 			public OutputKindValue11 OutputKind11;
 			public sbyte Output11;
 			public byte[] Unknown51;
-			public float Unknown52;
-			public float Unknown53;
-			public float Unknown54;
-			public float Unknown55;
-			public float Unknown56;
+			public uint Unknown52;
+			public uint Unknown53;
+			public uint Unknown54;
+			public uint Unknown55;
+			public uint Unknown56;
 			public List<UnknownBlock2> Unknown57;
 			public List<CompiledFunction> CompiledFunctions;
 			public List<CompiledColorFunction> CompiledColorFunctions;
@@ -138,7 +138,7 @@ namespace HaloOnlineTagTool.TagStructures
 				public StringId MaterialType;
 				public int Unknown;
 				public HaloTag Bitmap;
-				public float Unknown2;
+				public uint Unknown2;
 				public int Unknown3;
 				public short Unknown4;
 				public short Unknown5;
@@ -146,7 +146,7 @@ namespace HaloOnlineTagTool.TagStructures
 				public short Unknown7;
 				public short Unknown8;
 				public short Unknown9;
-				public float Unknown10;
+				public uint Unknown10;
 				public List<Function> Functions;
 
 				[TagStructure(Size = 0x24)]
@@ -154,8 +154,8 @@ namespace HaloOnlineTagTool.TagStructures
 				{
 					public int Unknown;
 					public StringId Name;
-					public float Unknown2;
-					public float Unknown3;
+					public uint Unknown2;
+					public uint Unknown3;
 					public byte[] Function2;
 				}
 			}
@@ -167,14 +167,14 @@ namespace HaloOnlineTagTool.TagStructures
 				public List<ShaderMap> ShaderMaps;
 				public List<Argument> Arguments;
 				public List<UnknownBlock> Unknown;
-				public float Unknown2;
+				public uint Unknown2;
 				public List<UnknownBlock2> Unknown3;
 				public List<UnknownBlock3> Unknown4;
 				public List<UnknownBlock4> Unknown5;
 				public List<Function> Functions;
 				public int Unknown6;
 				public int Unknown7;
-				public float Unknown8;
+				public uint Unknown8;
 				public short Unknown9;
 				public short Unknown10;
 				public short Unknown11;
@@ -210,7 +210,7 @@ namespace HaloOnlineTagTool.TagStructures
 				[TagStructure(Size = 0x4)]
 				public class UnknownBlock
 				{
-					public float Unknown;
+					public uint Unknown;
 				}
 
 				[TagStructure(Size = 0x2)]
@@ -222,7 +222,7 @@ namespace HaloOnlineTagTool.TagStructures
 				[TagStructure(Size = 0x6)]
 				public class UnknownBlock3
 				{
-					public float Unknown;
+					public uint Unknown;
 					public sbyte Unknown2;
 					public sbyte Unknown3;
 				}
@@ -239,8 +239,8 @@ namespace HaloOnlineTagTool.TagStructures
 				{
 					public int Unknown;
 					public StringId Name;
-					public float Unknown2;
-					public float Unknown3;
+					public uint Unknown2;
+					public uint Unknown3;
 					public byte[] Function2;
 				}
 			}
@@ -325,31 +325,31 @@ namespace HaloOnlineTagTool.TagStructures
 			[TagStructure(Size = 0x10)]
 			public class UnknownBlock2
 			{
-				public float Unknown;
-				public float Unknown2;
-				public float Unknown3;
-				public float Unknown4;
+				public uint Unknown;
+				public uint Unknown2;
+				public uint Unknown3;
+				public uint Unknown4;
 			}
 
 			[TagStructure(Size = 0x40)]
 			public class CompiledFunction
 			{
-				public float Unknown;
-				public float Unknown2;
-				public float Unknown3;
-				public float Unknown4;
-				public float Unknown5;
-				public float Unknown6;
-				public float Unknown7;
-				public float Unknown8;
-				public float Unknown9;
-				public float Unknown10;
-				public float Unknown11;
-				public float Unknown12;
-				public float Unknown13;
-				public float Unknown14;
-				public float Unknown15;
-				public float Unknown16;
+				public uint Unknown;
+				public uint Unknown2;
+				public uint Unknown3;
+				public uint Unknown4;
+				public uint Unknown5;
+				public uint Unknown6;
+				public uint Unknown7;
+				public uint Unknown8;
+				public uint Unknown9;
+				public uint Unknown10;
+				public uint Unknown11;
+				public uint Unknown12;
+				public uint Unknown13;
+				public uint Unknown14;
+				public uint Unknown15;
+				public uint Unknown16;
 			}
 
 			[TagStructure(Size = 0x10)]
