+++ HaloOnlineTagTool/TagStructures/Cloth.cs
@@ -12,43 +12,43 @@ namespace HaloOnlineTagTool.TagStructures
 	[TagStructure(Class = "clwd", Size = 0x94)]
 	public class Cloth
 	{
-		public float Unknown;
-		public float Unknown2;
-		public float Unknown3;
+		public uint Unknown;
+		public uint Unknown2;
+		public uint Unknown3;
 		public HaloTag Unknown4;
-		public float Unknown5;
-		public float Unknown6;
-		public float Unknown7;
-		public float Unknown8;
-		public float Unknown9;
-		public float Unknown10;
-		public float Unknown11;
-		public float Unknown12;
-		public float Unknown13;
-		public float Unknown14;
-		public float Unknown15;
-		public float Unknown16;
-		public float Unknown17;
-		public float Unknown18;
-		public float Unknown19;
-		public float Unknown20;
-		public float Unknown21;
-		public float Unknown22;
+		public uint Unknown5;
+		public uint Unknown6;
+		public uint Unknown7;
+		public uint Unknown8;
+		public uint Unknown9;
+		public uint Unknown10;
+		public uint Unknown11;
+		public uint Unknown12;
+		public uint Unknown13;
+		public uint Unknown14;
+		public uint Unknown15;
+		public uint Unknown16;
+		public uint Unknown17;
+		public uint Unknown18;
+		public uint Unknown19;
+		public uint Unknown20;
+		public uint Unknown21;
+		public uint Unknown22;
 		public List<UnknownBlock> Unknown23;
 		public List<UnknownBlock2> Unknown24;
-		public float Unknown25;
-		public float Unknown26;
-		public float Unknown27;
+		public uint Unknown25;
+		public uint Unknown26;
+		public uint Unknown27;
 		public List<UnknownBlock3> Unknown28;
 
 		[TagStructure(Size = 0x14)]
 		public class UnknownBlock
 		{
-			public float Unknown;
-			public float Unknown2;
-			public float Unknown3;
-			public float Unknown4;
-			public float Unknown5;
+			public uint Unknown;
+			public uint Unknown2;
+			public uint Unknown3;
+			public uint Unknown4;
+			public uint Unknown5;
 		}
 
 		[TagStructure(Size = 0x2)]
@@ -60,8 +60,8 @@ namespace HaloOnlineTagTool.TagStructures
 		[TagStructure(Size = 0x8)]
 		public class UnknownBlock3
 		{
-			public float Unknown;
-			public float Unknown2;
+			public uint Unknown;
+			public uint Unknown2;
 		}
 	}
 }
